getSpell, fix memory leak in VCS.cpp
In VCS.cpp, there was a memory use issue -- previously, we used temp = new, and delete temp at the end of the file--but, as it turns out, this caused an issue because the object pointers in the character class were pointing to objects in memory that no longer existed. Now, you can use the VCS functions to load an instance of character, characterclass, and race from a vcs, rvr, and rvc file. Additionally, a linked RVO is created with VCS files for character inventory.